Added utility function which gets and fixes image orientation
